+++ CHANGELOG.md
@@ -4,6 +4,7 @@
    - Added support for alternate rendering modes
  - Made force update override shouldComponentUpdate
    - Eliminates the need for `react-key` on root element
+- Fixed bug with dynamic locale loading UI refresh (timing issue)
 
 2.0.1
 -----